ver.2

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,7 +2,8 @@ package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
+import ru.yandex.practicum.filmorate.exception.NotFoundFilmException;
+import ru.yandex.practicum.filmorate.exception.NotFoundUserException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
@@ -10,8 +11,6 @@ import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.util.Collection;
 import java.util.stream.Collectors;
-
-import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 @Service
 public class FilmService {
@@ -30,10 +29,19 @@ public class FilmService {
     }
 
     public Film updateFilm(Film film) {
+        if (filmStorage.getFilmById(film.getId()) == null) {
+            throw new NotFoundFilmException("Фильм с ID=" + film.getId() + " не найден!");
+        }
         return filmStorage.updateFilm(film);
     }
 
     public Film deleteFilm(Integer id) {
+        if (id == null) {
+            throw new ValidationException("Передан пустой ID!");
+        }
+        if (filmStorage.getFilmById(id) == null) {
+            throw new NotFoundFilmException("Фильм с ID=" + id + " не найден!");
+        }
         return filmStorage.deleteFilm(id);
     }
 
@@ -42,6 +50,9 @@ public class FilmService {
     }
 
     public Film getFilmById(Integer id) {
+        if (filmStorage.getFilmById(id) == null) {
+            throw new NotFoundFilmException("Фильм с ID=" + id + " не найден!");
+        }
         return filmStorage.getFilmById(id);
     }
 
@@ -56,7 +67,7 @@ public class FilmService {
         if (userStorage.getUserById(userId) != null) {
             filmById.getLikes().add(userId);
         } else {
-            throw new ResponseStatusException(NOT_FOUND, "Пользователь c ID=" + userId + " не найден!");
+            throw new NotFoundUserException("Пользователь c ID=" + userId + " не найден!");
         }
     }
 
@@ -69,7 +80,7 @@ public class FilmService {
     public void deleteLike(Integer filmId, Integer userId) {
         Film filmById = filmStorage.getFilmById(filmId);
         if (!filmById.getLikes().remove(userId)) {
-            throw new ResponseStatusException(NOT_FOUND, "Лайк от пользователя c ID=" + userId + " не найден!");
+            throw new NotFoundUserException("Лайк от пользователя c ID=" + userId + " не найден!");
         }
     }
